+++ NEWS.md
@@ -49,6 +49,7 @@
 - Add a `purge` argument to `clean()` to optionally remove all target-level information.
 - Add a `namespace` argument to `cached()` so users can inspect individual `storr` namespaces.
 - Change `verbose` to numeric: 0 = print nothing, 1 = print progress on imports only, 2 = print everything.
+- Add a new `next_stage()` function to report the targets to be made in the next parallelizable stage.
 
 # Changes in release 4.4.0